requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-socketio==5.10.0
sqlalchemy==2.0.23
alembic==1.12.1
pillow==10.1.0
reportlab==4.0.7
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
python-dotenv==1.0.0
redis==5.0.1
psycopg2-binary==2.9.9
django-environ==0.11.2 